requirements.txt
numpy
tensorflow
google-cloud-aiplatform
streamlit==1.13.0
gunicorn